% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_velocidade.R
\name{calcular_velocidade}
\alias{calcular_velocidade}
\title{Calcular velocidade média de caminhada}
\usage{
calcular_velocidade(distancia, tempo)
}
\arguments{
\item{distancia}{Distância percorrida (cm)}

\item{tempo}{Tempo (s)}
}
\value{
Velocidade média (cm/s)
}
\description{
Função genérica para calcular velocidade média (cm/s)
}
\examples{
calcular_velocidade(10, 5)

}
