% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar_mon_ex.R
\docType{data}
\name{dar_mon_ex}
\alias{dar_mon_ex}
\title{Example data frame for Days in AR calculation}
\format{
A data frame with 12 rows and 3 variables
#' \describe{
\item{date}{column containing first day of month}
\item{gct}{column containing the month's total Gross Charges}
\item{earb}{column containing month's Ending AR balance}
...
}
}
\usage{
dar_mon_ex
}
\description{
Example of a data frame needed for the
Days in AR calculation
}
\keyword{datasets}
