% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar_month.R
\name{dar_month}
\alias{dar_month}
\title{Calculate Monthly Days in AR}
\usage{
dar_month(df, date = date, gct = gct, earb = earb, dart = 35)
}
\arguments{
\item{df}{data frame containing at least three columns:
a date column, a gross charges column and an ending AR column}

\item{date}{column containing a date within the month that
Days in AR is to be calculated}

\item{gct}{column containing a month's total Gross Charges}

\item{earb}{column containing a month's Ending AR balance}

\item{dart}{target Days in AR, default is 35 days}
}
\value{
data frame
}
\description{
Calculate Monthly Days in AR
}
\examples{
dar_month(dar_mon_ex, date, gct, earb, dart = 39.445)
}
