% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar_qtr.R
\name{dar_qtr}
\alias{dar_qtr}
\title{Calculate Quarterly Days in AR}
\usage{
dar_qtr(df, date = date, gct = gct, earb = earb, dart = 35)
}
\arguments{
\item{df}{data frame containing at least three columns:
a date column, a gross charges column and an ending AR column}

\item{date}{column containing a date within the month that
Days in AR is to be calculated}

\item{gct}{column containing a month's total Gross Charges}

\item{earb}{column containing a month's Ending AR balance}

\item{dart}{target Days in AR, default is 35 days}
}
\value{
data frame
}
\description{
Calculate Quarterly Days in AR
}
\examples{
dar_mon_ex <- data.frame(
date = as.Date(c(
"2022-01-01", "2022-02-01", "2022-03-01",
"2022-04-01", "2022-05-01", "2022-06-01",
"2022-07-01", "2022-08-01", "2022-09-01",
"2022-10-01", "2022-11-01", "2022-12-01")),

gct = c(
325982.23, 297731.74, 198655.14,
186047.56, 123654.34, 131440.28,
153991.95, 156975.52, 146878.12,
163799.44, 151410.74, 169094.46),

earb = c(
288432.52, 307871.08, 253976.56,
183684.92, 204227.59, 203460.47,
182771.32, 169633.64, 179347.72,
178051.11, 162757.49, 199849.32))

dar_qtr(dar_mon_ex, date, gct, earb, dart = 40)
}
