% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging.R
\name{days_between}
\alias{days_between}
\title{Calculate Number of Days Between Two Dates}
\usage{
days_between(df, from, to = NULL)
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>}}

\item{from}{\verb{[character]} column of start dates}

\item{to}{\verb{[character]} column of end dates}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Calculate Number of Days Between Two Dates
}
\examples{
generate_data(10)[c(
  "date_service",
  "charges",
  "payer")] |>
  days_between(date_service)

}
