% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth_from_randomForest}
\alias{depth_from_randomForest}
\title{depth_from_randomForest}
\usage{
depth_from_randomForest(treelike)
}
\arguments{
\item{treelike}{a treelike structure}
}
\description{
depth_from_randomForest
}
