% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthAddProcureCost}
\alias{dietBreadthAddProcureCost}
\title{Diet breadth add procurement cost}
\usage{
dietBreadthAddProcureCost(u0, h, procure_cost)
}
\arguments{
\item{u0}{Vector of base resource utilities}

\item{h}{Vector of resource handling times}

\item{procure_cost}{Vector of resource procurement costs (as a rate)}
}
\value{
Vector of resource utilities accounting for procurement cost
}
\description{
Diet breadth add procurement cost
}
