% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthAnalysis}
\alias{dietBreadthAnalysis}
\title{Diet breadth summary analysis}
\usage{
dietBreadthAnalysis(u, h, lambda, c_s = 0, resNames = NA)
}
\arguments{
\item{u}{Vector of resource utilities}

\item{h}{Vector of resource handling times}

\item{lambda}{Vector of encounter rates}

\item{c_s}{Loss of utility per unit time while searching (default 0)}

\item{resNames}{Vector of resource names (default none)}
}
\description{
Diet breadth summary analysis
}
