% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthDensityToEncounterRate}
\alias{dietBreadthDensityToEncounterRate}
\title{Diet breadth density to encounter rate}
\usage{
dietBreadthDensityToEncounterRate(density, search_radius, search_speed)
}
\arguments{
\item{density}{Density of resource on landscape (Number / Area)}

\item{search_radius}{Visibility of resource to forager (Length)}

\item{search_speed}{Speed at which forager moves through the landscape while searching (Length / Time)}
}
\value{
Encounter rate with the resource (Number / Time)
}
\description{
Diet breadth density to encounter rate
}
