% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthOptimalReturnRate}
\alias{dietBreadthOptimalReturnRate}
\title{Diet breadth optimal return rate}
\usage{
dietBreadthOptimalReturnRate(v, h, lambda, c_s_dot = 0)
}
\arguments{
\item{v}{Vector of resource values}

\item{h}{Vector of resource handling times}

\item{lambda}{Vector of encounter rates}

\item{c_s_dot}{Loss of value per unit time while searching (default 0)}
}
\value{
A boolean vector indicating whether each resource is in the optimal set
}
\description{
Diet breadth optimal return rate
}
