% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthOptimalSet}
\alias{dietBreadthOptimalSet}
\title{Diet breadth model}
\usage{
dietBreadthOptimalSet(v, h, lambda, c_s_dot = 0)
}
\arguments{
\item{v}{Vector of resource values}

\item{h}{Vector of resource handling times}

\item{lambda}{Vector of encounter rates}

\item{c_s_dot}{Loss of value per unit time while searching (default 0)}
}
\value{
A boolean vector indicating whether each resource is in the optimal set
}
\description{
\itemize{
  \item{dietBreadthOptimalSet determines the optimal set of resources to take on encounter}
  \item{dietBreadthReturnRate calculates the return rate (kcal/hr) given a set of items in the diet.}
  \item{dietBreadthOptimReturnRate calculates return rate for the optimal set.}
}
}
\details{
A forager searches the landscape for resources to procure. Each resource is encountered randomly as a Poisson process with encounter rate lambda_i. On encountering a resource, the forager chooses whether or not to procure it. If not, the forager continues searching. If so, the forager pursues and processes it with handling time h_i, receiving a value from the resource of v_i. Usually the value is the food energy provided by the item. Optionally, search is costly to the forager at the rate c_s_dot. That is, c_s_dot is the rate of loss of value (usually energy / time).

The overall return rate of foraging is (using R syntax):

returnRate = (-c_s_dot + sum(lambda * v)) / (1 + sum(lambda * h))

The optimal diet is the set of items that maximizes this returnRate. Define

rank_i = v_i / h_i

as the rank (or on encounter return rate) of resource i. To determine the optimal set, order resources by rank from best to worst. Add resources to the diet in this rank order until the return rate no longer increases from adding additional items. This occurs when the on encounter return rate of a new item in the diet is lower than the overall return rate from procuring all the items that are more highly ranked.
}
\references{
MacArthur and Pianka 1966 -- On Optimal Use of a Patchy Environment. American Naturalist. 100:603-609.
}
