% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthReturnRate}
\alias{dietBreadthReturnRate}
\title{Diet breadth return rate}
\usage{
dietBreadthReturnRate(v, h, lambda, c_s_dot = 0, include = NA)
}
\arguments{
\item{v}{Vector of resource values}

\item{h}{Vector of resource handling times}

\item{lambda}{Vector of encounter rates}

\item{c_s_dot}{Loss of value per unit time while searching (default 0)}
}
\value{
A boolean vector indicating whether each resource is in the optimal set
}
\description{
Diet breadth return rate
}
