% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{Distance matrix from unsupervised random forest.}
\usage{
dist(dataset, method = "synthetic", prop = TRUE,
  predictMethod = "terminalNodes")
}
\arguments{
\item{dataset}{A dataframe}

\item{method}{(string) See 'method' in synthetic_forest}

\item{prop}{(flag) See 'prop' in synthetic_forest}

\item{predictMethod}{See 'method' in forage}
}
\description{
Distance matrix from unsupervised random forest.
}
