% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extract.R
\name{distExtractPair}
\alias{distExtractPair}
\title{distExtractPair}
\usage{
distExtractPair(object, index1, index2)
}
\arguments{
\item{object}{distance object}

\item{index1}{index}

\item{index2}{index}
}
\value{
(A number) Distance between index1 and index2.
}
\description{
distExtractPair
}
