% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ooberror.R
\name{extract_ooberror.randomForest}
\alias{extract_ooberror.randomForest}
\title{Extract OOB error from a randomForest object}
\usage{
\method{extract_ooberror}{randomForest}(object)
}
\arguments{
\item{object}{forest object of class: randomForest}
}
\description{
Extract OOB error from a randomForest object
}
