% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ooberror.R
\name{extract_ooberror.ranger}
\alias{extract_ooberror.ranger}
\title{Extract OOB error from a ranger object}
\usage{
\method{extract_ooberror}{ranger}(object)
}
\arguments{
\item{object}{forest object of class: ranger}
}
\description{
Extract OOB error from a ranger object
}
