% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ooberror.R
\name{extract_ooberror}
\alias{extract_ooberror}
\title{Extract OOB error from a forest object}
\usage{
extract_ooberror(object)
}
\arguments{
\item{object}{forest object of class: ranger, randomForest}
}
\value{
A numerical value with an attribute. See description.
}
\description{
Extract Out of box error from a forest object. Supports objects
  of these classes: ranger, randomForest. The result adds an attribute
  'measure' indicating what the result is measuring.
}
\examples{
temp <- ranger::ranger(Species ~., data = iris)
res <- extract_ooberror(temp)
res
attr(res, "measure")

temp <- ranger::ranger(Sepal.Length ~., data = iris)
res <- extract_ooberror(temp)
res
attr(res, "measure")

temp <- randomForest::randomForest(Species ~., data = iris)
res <- extract_ooberror(temp)
res
attr(res, "measure")

temp <- randomForest::randomForest(Sepal.Length ~., data = iris)
res <- extract_ooberror(temp)
res
attr(res, "measure")

}
