% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forager-package.R
\docType{package}
\name{forager-package}
\alias{forager}
\alias{forager-package}
\title{forager: Healthcare Revenue Cycle Analysis}
\description{
The goal of forager is to provide a suite of analytic tools for healthcare revenue cycle management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/forager}
  \item \url{https://andrewallenbruce.github.io/forager/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/forager/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
