% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data}
\alias{generate_data}
\title{Generate mock coding/billing data frame}
\usage{
generate_data(rows = 100)
}
\arguments{
\item{rows}{number of rows to generate; default is 100}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate mock coding/billing data frame
}
\examples{
generate_data(rows = 5)

}
