% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_impute.R
\name{impute_stack}
\alias{impute_stack}
\title{Unsupervised imputation by stacking complete and incomplete data}
\usage{
impute_stack(dataset, newdata, method = "missforest", seed = 1L, ...)
}
\arguments{
\item{dataset}{dataset}

\item{newdata}{newdata}

\item{method}{One among: 'mf', 'proximity'}

\item{seed}{seed}

\item{...}{dotdotdot}
}
\value{
completed dataset
}
\description{
impute_stack
}
