% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_ex}
\alias{load_ex}
\title{Load example data}
\usage{
load_ex(name = "aging_ex")
}
\arguments{
\item{name}{\verb{<chr>} name of example dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Load example data
}
\keyword{internal}
