% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R
\name{mock_claims}
\alias{mock_claims}
\title{Generate mock coding/billing data frame}
\usage{
mock_claims(rows = 1000L, days = FALSE)
}
\arguments{
\item{rows}{\verb{<int>} number of rows to generate}

\item{days}{\verb{<lgl>} add counts of days between events}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate mock coding/billing data frame
}
\examples{
mock_claims(rows = 100)
}
