% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_generators.R
\name{mock_parbx}
\alias{mock_parbx}
\title{Generate Mock PARBx data}
\usage{
mock_parbx(payers = payer_names(), ...)
}
\arguments{
\item{payers}{\verb{[character]} vector of payer names}

\item{...}{\verb{[dots]} additional arguments}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate Mock PARBx data
}
\examples{
# Every name generates 60 rows of data
mock_parbx()

}
