% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ooberror.randomForest}
\alias{ooberror.randomForest}
\alias{oob_error.randomForest}
\title{Extract OOB error from a randomForest model}
\usage{
\method{oob_error}{randomForest}(model)
}
\arguments{
\item{model}{forest model of class: randomForest}
}
\description{
Extract OOB error from a randomForest model
}
