% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ooberror.ranger}
\alias{ooberror.ranger}
\alias{oob_error.ranger}
\title{Extract OOB error from a ranger model}
\usage{
\method{oob_error}{ranger}(model)
}
\arguments{
\item{model}{forest model of class: ranger}
}
\description{
Extract OOB error from a ranger model
}
