% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\name{perda_relativa}
\alias{perda_relativa}
\title{Calcula perda relativa a partir do ganho percentual}
\usage{
perda_relativa(ganho_percentual)
}
\arguments{
\item{ganho_percentual}{número ou vetor com o ganho percentual}
}
\value{
perda percentual correspondente
}
\description{
Converte um aumento percentual (ganho) em perda relativa equivalente.
}
\examples{
perda_relativa(240)
perda_relativa(c(50, 100, 200))
}
