% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{predict_depth.randomForest}
\alias{predict_depth.randomForest}
\title{predict_depth.randomForest}
\usage{
\method{predict_depth}{randomForest}(object, data, nproc = 1)
}
\arguments{
\item{object}{A tree ensemble model}

\item{data}{data}

\item{nproc}{number of cores to use for parallel processing}
}
\description{
predict_depth.randomForest
}
