% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{predict_depth}
\alias{predict_depth}
\title{predict_depth}
\usage{
predict_depth(object, data, parallel = TRUE)
}
\arguments{
\item{object}{A tree ensemble model}

\item{data}{data}

\item{parallel}{(flag) Whether to use multicore parallel processing on
unix-alike systems.}
}
\description{
predict_depth
}
