% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{predict_depth_observations_terminalNodes}
\alias{predict_depth_observations_terminalNodes}
\title{predict_depth_observations_terminalNodes}
\usage{
predict_depth_observations_terminalNodes(object, data, parallel = TRUE)
}
\description{
predict_depth_observations_terminalNodes
}
