% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_net.R
\name{predict_net}
\alias{predict_net}
\title{Lirov's Net Payment Estimation}
\usage{
predict_net(df, date, gct, earb, net, parb_120)
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>}}

\item{date}{column of \verb{<date>}s}

\item{gct}{\verb{<dbl>} column of total Gross Charges}

\item{net}{column of net payments}

\item{parb_120}{column of percentage of AR beyond 120 days old}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Lirov's Net Payment Estimation
}
\examples{
predict_net(
  net_ex(),
  date,
  gct,
  earb,
  net,
  parb_120)

}
