% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_compute.R
\name{predict_outlyingness_observations_terminalNodes}
\alias{predict_outlyingness_observations_terminalNodes}
\title{predict_outlyingness_observations_terminalNodes}
\usage{
predict_outlyingness_observations_terminalNodes(object, newdata, classes)
}
\description{
predict_outlyingness_observations_terminalNodes
}
