% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalNodes.R
\name{predict_terminalNodesMatrix_terminalNodes}
\alias{predict_terminalNodesMatrix_terminalNodes}
\title{predict_terminalNodesMatrix_terminalNodes}
\usage{
predict_terminalNodesMatrix_terminalNodes(object, newdata)
}
\arguments{
\item{object}{object}

\item{newdata}{newdata}
}
\description{
predict_terminalNodesMatrix_terminalNodes
}
