% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_claims.R
\name{prep_claims}
\alias{prep_claims}
\title{Prep mock coding/billing data frame}
\usage{
prep_claims(df, ...)
}
\arguments{
\item{df}{\verb{[data.frame]} data frame generated by \code{mock_claims()}}

\item{...}{\verb{[dots]} additional arguments}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Prep mock coding/billing data frame
}
\examples{
prep_claims(mock_claims(rows = 5))

}
