% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\name{regra3}
\alias{regra3}
\title{Regra de 3 simples}
\usage{
regra3(a, b, y)
}
\arguments{
\item{a}{valor a}

\item{b}{valor b}

\item{y}{valor y (não pode ser zero)}
}
\value{
resultado da regra de 3
}
\description{
Regra de 3 simples
}
