% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_compute.R
\name{rf_compute}
\alias{rf_compute}
\title{Obtain various outputs using a trained random forest or extremely
  randomized trees on new data.}
\usage{
rf_compute(object, newdata, what = "dissimilarity",
  method = "terminalNodes", context = "observations", classes = NULL, ...)
}
\arguments{
\item{object}{Trained random forest model. Currently, only trained random
forests from 'ranger' and 'randomForest' are supported.}

\item{newdata}{(object inheriting 'data.frame' class) A dataframe.}

\item{what}{(string) Type of output. Following are implemented:
terminalNodesMatrix, dissimilarity, proximity, outlyingness. Default is
'dissimilarity'.}

\item{method}{(string) Method to obtain the output. Following are
implemented: terminalNodes. Default is 'terminalNodes'.}

\item{classes}{(factor) Required when 'what' is 'outlyingness'.}
}
\value{
The following are returned depending on 'what': \itemize{ \item
  \strong{terminalNodesMatrix}: A matrix with dimension number of
  observations times the number of trees. An entry is ID of the terminal
  node. \item \strong{dissimilarity}: A 'dist' object with size of number of
  observations. \item \strong{proximity}: A 'simil/dist' object(from 'proxy'
  package) with size of number of observations. \item \strong{outlyingness}:
  A vector of measure of outlyingness of each observation from its class. }
}
\description{
Obtain terminalNodesMatrix, dissimilarity, proximity,
  outlyingness on new data using a trained random forest. Currently, trained
  random forests from 'ranger' and 'randomForest' packages are supported. See
  details for the explanation about various outputs.
}
\details{
TODO
Add details
comments please
Set links in expected style
Provide examples
Add references
}
