% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_synthetic.R
\name{rf_synthetic}
\alias{rf_synthetic}
\title{Grow a random forest (or extratrees) on synthetic data.}
\usage{
rf_synthetic(dataset, prop = TRUE, seed = 1L, ...)
}
\arguments{
\item{dataset}{(object of class 'data.frame') A dataframe}

\item{prop}{(flag) Random sampling of covariates (when prop = TRUE) to
generate synthetic data. Else, uniform sampling is used.}

\item{seed}{(a positive integer) Seed for sampling.}

\item{...}{Arguments to be passed to \code{\link[ranger]{ranger}}.}
}
\value{
A random forest model of class 'ranger'.
}
\description{
Builds a random forest model to classify actual vs synthetic
  data where synthetic data is created by sampling each covariate as
  suggested in
  \link[=https://www.stat.berkeley.edu/~breiman/Using_random_forests_v4.0.pdf]{Understaning
   random forests} by Brieman.
}
\details{
TODO
Provide examples


\link[=https://www.stat.berkeley.edu/~breiman/Using_random_forests_v4.0.pdf]{Understaning
random forests} by Brieman involves creating synthetic data by sampling
randomly from unvariate distributions of each covariate(feature). This
supports two methods: First, where proportions or distribution is taken into
account when sampling at random, second where the data is sampled assuming
uniform distribution. The former corresponds to "Addcl1" from Horvath's
\link[=https://doi.org/10.1198/106186006X94072]{paper} and latter corresponds
to "addc2". A random forest model is built using \pkg{ranger} to learn what
separates the actual data from the synthetic data. Default value of number of
trees grown is 1e3 and  minimum terminal node size is set to 5.
}
\references{
\itemize{

\item \link[=https://doi.org/10.1198/106186006X94072]{Unsupervised Learning
With Random Forest Predictors} by Tao Shi & Steve Horvath.

\item
\link[=https://www.stat.berkeley.edu/~breiman/Using_random_forests_v4.0.pdf]{Understaning
random forests} by Brieman. }
}
