% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughImpute.R
\name{roughImpute}
\alias{roughImpute}
\title{Unsupervised median/mode imputation for dataframes}
\usage{
roughImpute(dataset)
}
\arguments{
\item{dataset}{A dataframe}
}
\value{
A dataframe (also inherits data.table class)
}
\description{
Imputes missing values integer/numeric columns with their
  median. Imputes missing calues in character/factor columns with their mode.
  If a column does not contain even one non-missing value, then error is
  thrown.
}
\examples{
iris_missing <- missRanger::generateNA(iris, p = 0.2, seed = 1)
roughImpute(iris_missing)
}
