% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Sample a categorical or numeric vector using random or uniform sampling.}
\usage{
sampler(x, prop, seed)
}
\arguments{
\item{x}{The vector.}

\item{prop}{Random sampling when prop = TRUE. Else, uniform sampling is used.}

\item{seed}{Seed for sampling.}
}
\value{
A sampled vector of the same length.
}
\description{
This is intended to be used for synthetic data generation.
}
