% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular_forrageamento.R
\name{simular_forrageamento}
\alias{simular_forrageamento}
\title{Simula comportamento de forrageamento}
\usage{
simular_forrageamento(n = 10, taxa_saida = 0.3)
}
\arguments{
\item{n}{Numero de abelhas}

\item{taxa_saida}{Taxa de saida de abelhas}
}
\value{
Data frame com tempos simulados
}
\description{
Simula comportamento de forrageamento
}
