% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_claims.R
\name{summarise_claims}
\alias{summarise_claims}
\title{Summarise mock coding/billing data frame}
\usage{
summarise_claims(df, vars = c(dplyr::starts_with("days_"), dar))
}
\arguments{
\item{df}{\verb{[data.frame]} data generated from \code{prep_claims(mock_claims())}}

\item{vars}{\verb{[character]} variables to summarise, e.g. \code{c(days_srv, days_rec)}}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Summarise mock coding/billing data frame
}
\examples{
x <- mock_claims(5000)
x

x <- prep_claims(x)
x

summarise_claims(x) |>
dplyr::glimpse()

dplyr::group_by(x,
year = ymd::year(date_srv),
month = ymd::month(date_srv)) |>
summarise_claims() |>
dplyr::glimpse()
}
