% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_forest.R
\name{synthetic_forest}
\alias{synthetic_forest}
\title{Grow a tree ensemble on synthetic data}
\usage{
synthetic_forest(dataset, prop = TRUE, seed = 1L,
  implementation = "ranger", ...)
}
\arguments{
\item{dataset}{A dataframe}

\item{prop}{(flag) Random sampling of covariates (when prop = TRUE) to
generate synthetic data. Else, uniform sampling is used.}

\item{seed}{(a positive integer) Seed for sampling.}

\item{implementation}{(string) Implemenation to use to build the model. The
following are supported: 'ranger', 'randomForest'.}

\item{...}{Arguments to be passed to implementation.}
}
\value{
A tree ensemble with one these classes: 'ranger', 'randomForest'
}
\description{
Builds a random forest model to classify actual vs synthetic
  data where synthetic data is created by sampling each covariate as
  suggested in
  \href{https://www.stat.berkeley.edu/~breiman/Using_random_forests_v4.0.pdf}{Understanding
   random forests} by Brieman.
}
\details{
\href{https://www.stat.berkeley.edu/~breiman/Using_random_forests_v4.0.pdf}{Understanding
random forests} by Brieman involves creating synthetic data by sampling
randomly from unvariate distributions of each covariate(feature). This
supports two methods: First, where proportions or distribution is taken into
account when sampling at random, second where the data is sampled assuming
uniform distribution. The former corresponds to "Addcl1" from Horvath's
\href{https://doi.org/10.1198/106186006X94072}{paper} and latter corresponds
to "addc2". A random forest model is built using \pkg{ranger} or
\pkg{randomForest} to learn what separates the actual data from the synthetic
data. Default value of number of trees grown is 1000 and  minimum node size
to split is set to 5.
}
\examples{

# ranger
model_ranger <- synthetic_forest(iris, implementation = "ranger")
oob_error(model_ranger)

# randomForest
model_rf <- synthetic_forest(iris, implementation = "randomForest")
oob_error(model_rf)

# extratrees
model_et <- synthetic_forest(iris, implementation = "ranger", splitrule = "extratrees")
oob_error(model_et)

}
\references{
\itemize{

  \item \href{https://doi.org/10.1198/106186006X94072}{Unsupervised Learning
  With Random Forest Predictors} by Tao Shi & Steve Horvath.

  \item
  \href{https://www.stat.berkeley.edu/~breiman/Using_random_forests_v4.0.pdf}{Understanding
   random forests} by Brieman.

  }
}
