% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\title{Run unsupervised random forest}
\usage{
unsupervised(dataset, method = "synthetic", seed = 1L, imputeArgs = NULL,
  ...)
}
\description{
Implements various unsupervised random forest methods. See details.
}
\details{
Currently, the "terminalnode" method is implemented. This was suggested by Brieman (https://www.stat.berkeley.edu/~breiman/RandomForests/cc_home.htm#unsup) and two different sampling methods were suggested by
}
