% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forager_package.R
\docType{package}
\name{unsupervised_package}
\alias{forager}
\alias{forager-package}
\alias{unsupervised_package-package}
\title{Unsupervised random forest methods.}
\description{
Provides an extensible framework to compute various measures like proximity or dissimilarity between observations/trees, estimate outlyingness of an observation and an imputation technique is implemented under supervised and unsupervised setting. Currently, models from ranger and randomForest packages are supported.
}
\author{
\strong{Maintainer}: KS Srikanth \email{sri.teach@gmail.com}

}
