% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\name{velocidade_media}
\alias{velocidade_media}
\title{Velocidade média}
\usage{
velocidade_media(distancia, tempo)
}
\arguments{
\item{distancia}{vetor de distâncias}

\item{tempo}{vetor de tempos}
}
\value{
vetor de velocidades
}
\description{
Velocidade média
}
