% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar.R
\name{avg_dar}
\alias{avg_dar}
\title{Calculate Average Days in AR}
\usage{
avg_dar(df, date, gct, earb, dart = 35, period = c("month", "quarter"))
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>} with three required columns: date,
gross charges column and ending Accounts Receivables balance}

\item{date}{column of \verb{<date>}s}

\item{gct}{\verb{<dbl>} column of Gross Charges}

\item{earb}{\verb{<dbl>} column of Ending AR balances}

\item{dart}{\verb{<dbl>} Target Days in AR, default is \code{35} days}

\item{period}{\verb{<chr>} string specifying the calculation period; one of
\code{"month"}, \code{"quarter"}, or \code{"year"}; defaults to \code{"month"}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Calculate Average Days in AR
}
\examples{
avg_dar(df     = dar_ex(),
        date   = monthdate,
        gct    = gross_charges,
        earb   = ending_ar,
        dart   = 35,
        period = "month")

avg_dar(df     = dar_ex(),
        date   = monthdate,
        gct    = gross_charges,
        earb   = ending_ar,
        dart   = 35,
        period = "quarter")

}
