% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging.R
\name{bin_aging}
\alias{bin_aging}
\title{Apply 30-Day Aging Bins}
\usage{
bin_aging(df, date)
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>} with a required date column}

\item{date}{column of \verb{<date>}s}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Apply 30-Day Aging Bins
}
\examples{
binned <- bin_aging(load_ex(), dos)

head(binned)

binned |>
  dplyr::arrange(aging_bin) |>
  dplyr::summarise(
  n_claims = dplyr::n(),
  balance = sum(charges),
  .by = aging_bin)

binned |>
  dplyr::arrange(aging_bin) |>
  dplyr::summarise(
  n_claims = dplyr::n(),
  balance = sum(charges),
  .by = c(aging_bin, ins_name))

}
