% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging.R
\name{bin_aging}
\alias{bin_aging}
\title{Apply 30-Day Aging Bins}
\usage{
bin_aging(df, ndays, bin_type = c("case", "chop"))
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>}}

\item{ndays}{\verb{<dbl>} column of counts of days elapsed to bin by}

\item{bin_type}{\verb{<chr>} string specifying the bin type; one of "chop", "cut" or "ivs"}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Apply 30-Day Aging Bins
}
\examples{
generate_data(10)[c(
  "date_srvc",
  "charges",
  "payer")] |>
  days_between(date_srvc) |>
  bin_aging(days_in_ar)

load_ex("aging_ex") |>
  dplyr::select(dos, charges, ins_name) |>
  days_between(dos) |>
  bin_aging(days_in_ar) |>
  dplyr::arrange(aging_bin) |>
  dplyr::group_by(
    year = clock::get_year(dos),
    month = clock::date_month_factor(dos),
  ) |>
  janitor::tabyl(ins_name, aging_bin, year)

}
