% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_breadth.R
\name{dietBreadthOptimalReturnRate}
\alias{dietBreadthOptimalReturnRate}
\title{Diet breadth optimal return rate}
\usage{
dietBreadthOptimalReturnRate(u, h, lambda, c_s = 0)
}
\arguments{
\item{u}{Vector of resource utilities}

\item{h}{Vector of resource handling times}

\item{lambda}{Vector of encounter rates}

\item{c_s}{Loss of utility per unit time while searching (default 0)}
}
\description{
Diet breadth optimal return rate
}
