% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{Distance matrix from unsupervised tree ensemble}
\usage{
dist(dataset, method = "synthetic", predictMethod = "terminalNodes", ...)
}
\arguments{
\item{dataset}{A dataframe}

\item{method}{(string) Method to build the tree ensemble. These are
supported: 'synthetic'}

\item{predictMethod}{Method to predict. These are supported: 'terminalNodes'}

\item{...}{Arguments for the tree ensembler}
}
\value{
A object of class 'dist'
}
\description{
Compute 'dist' object from unsupervised tree ensemble
}
\details{
If method is 'synthetic', a tree ensemble is grown to seperate
  actual data from synthetic data using 'synthetic_forest' function. If
  predictMethod is 'terminalNodes', distance matrix is computed from
  cooccurance in terminal nodes using 'forage' function.
}
\examples{
dm <- dist(iris[, 1:4])
attr(dm, "Size")
}
