% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extract.R
\name{dist_extract}
\alias{dist_extract}
\title{dist_extract}
\usage{
dist_extract(object, i, j, product = "inner")
}
\arguments{
\item{object}{distance object}

\item{i}{index}

\item{j}{index}

\item{product}{(string) product type. One among: 'inner', 'outer'.}
}
\value{
A vector of distances when 'inner' is used. A matrix of distances
  when 'outer' is used.
}
\description{
dist_extract
}
