% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage.R
\name{forage}
\alias{forage}
\title{Obtain various outputs using a trained tree ensembles on new data.}
\usage{
forage(object, newdata, what = "dissimilarity", method = "terminalNodes",
  context = "observations", classes = NULL, ...)
}
\arguments{
\item{object}{Object of class 'ranger' and 'randomForest'.}

\item{newdata}{(object inheriting 'data.frame' class) A dataframe.}

\item{what}{(string) Type of output. Following are implemented:
terminalNodesMatrix, dissimilarity, proximity, outlyingness, depth. Default
is 'dissimilarity'.}

\item{method}{(string) Method to obtain the output. Following are
implemented: terminalNodes. Default is 'terminalNodes'.}

\item{context}{(string) Specify whether output should be computed for
'observations' or 'trees'.}

\item{classes}{(factor) Required when 'what' is 'outlyingness'}

\item{...}{Currently not in use.}
}
\value{
The following are returned depending on 'what':

  \itemize{

  \item \strong{terminalNodesMatrix}: A matrix with dimension number of
  observations times the number of trees. An entry is ID of the terminal
  node.

  \item \strong{dissimilarity}: A 'dist' object with size of number of
  observations.

  \item \strong{proximity}: A 'simil/dist' object(from 'proxy' package) with
  size of number of observations.

  \item \strong{outlyingness}: A vector of measure of outlyingness of each
  observation from its class.

  \item \strong{depth}: A matrix with dimension number of observations times
  the number of trees. An entry is depth of tgher

  }
}
\description{
Obtain terminalNodesMatrix, dissimilarity, proximity,
  outlyingness on new data using a tree ensemble. Currently, ensembles from
  'ranger' and 'randomForest' packages are supported. See details for the
  explanation about various outputs.
}
\details{
TODO
}
