% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_rfe.R
\name{forest_rfe}
\alias{forest_rfe}
\title{lightweight implementation of RFE using ranger}
\usage{
forest_rfe(dataset, responseVarName, sizes, sampleprop = 0.2,
  nsamples = 10, seed = 1, ...)
}
\arguments{
\item{dataset}{(object inheriting data.frame class) A dataframe}

\item{responseVarName}{(string) Name of the response variable}

\item{sizes}{(integer vector) Vector of number of variables. When missing, sizes will be sequence of nc/2^i where the sequnce ranges from nc(number of columns) to 2.}

\item{sampleprop}{(A real number between 0 and 1 or a vector) Proportion of observations. If not a single number and sizes is specified, this vector should have same length as sizes.
per sample}

\item{nsamples}{(positive integer or a vector) Number of samples. If not a single number and sizes is specified, this vector should have same length as sizes.}

\item{seed}{(positive integer) Seed}

\item{...}{Arguments to be passed to \code{\link[ranger]{ranger}}}
}
\value{
A list with:

\itemize{

  \item (rfeTable) A tibble with three columns:

  \itemize{

    \item size: Number of variables used \item ooberror: Out-of-box error of
  the forest

    \item varimp: A list-column where each item is a data.frame with
  variable names and importance

    }

 \item (oobchangeTable) A dataframe with five columns sorted by absolute value of the variable 'oepc'.

 \itemize{

   \item variable: Name of the variable that got removed at some stage

   \item size: Number of variables that were considered before removing the variable

   \item reducedSize: Number of the variables at next stage. Gives an idea of how many variables were reduced at that stage.

   \item oepc: OOB error percentage change

   \item importance: Importance of the variable at the stage when the variable was decided to be removed.

  }
}
}
\description{
For datasets with large number of predictors, this
  implementation has these modifications to regular recursive feature
  elimination procedure:

  \itemize{

  \item Use oob prediction error as a proxy to model performance.

  \item Build forests \code{\link[ranger]{ranger}} on samples of data and
  average variable importance and oob prediction error.

  }

  For a comprehensive RFE procedure with resampling, use
  \code{\link[caret]{rfe}}
}
\examples{
temp <- forest_rfe(iris, "Species")
temp

temp <- forest_rfe(iris
                   , "Species"
                   , sizes = c(4,2)
                   , sampleprop = c(0.2, 0.3)
                   , nsamples = c(20, 30)
                   )
temp

temp <- forest_rfe(iris
                   , "Species"
                   , sizes = c(4,2)
                   , sampleprop = 0.1
                   , nsamples = c(20, 30)
                   )
temp

temp <- forest_rfe(iris
                   , "Species"
                   , sizes = c(4,2)
                   , sampleprop = c(0.2, 0.3)
                   , nsamples = 10
                   )
temp

temp <- forest_rfe(iris
                   , "Species"
                   , sizes = c(4,2)
                   , sampleprop = c(0.2, 0.3)
                   , nsamples = 10
                   , mtry = list(3, 2)
                   , num.trees = list(500, 1000)
                   , case.weights = replicate(2, runif(150), simplify = FALSE)
                   )
temp

}
\references{
\itemize{

  \item
  \href{https://topepo.github.io/caret/recursive-feature-elimination.html}{RFE
   using caret}

  \item
  \href{http://scikit-learn.org/stable/modules/generated/sklearn.feature_selection.RFE.html}{RFE
   using scikit learn}

  }
}
