% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_rfe.R
\name{forest_rfe}
\alias{forest_rfe}
\title{lightweight implementation of RFE using ranger random forest implementation}
\usage{
forest_rfe(dataset, responseVarName, sizes, sampleprop = 0.2, nsamples = 10,
  seed = 1, ...)
}
\arguments{
\item{dataset}{(object inheriting data.frame class) A dataframe}

\item{responseVarName}{(string) Name of the response variable}

\item{sizes}{(integer vector) Vector of number of variables}

\item{sampleprop}{(real number between 0 and 1) Proportion of observations per sample}

\item{nsamples}{(positive integer) Number of samples}

\item{seed}{(positive integer) Seed}

\item{...}{Arguments to be passed to `ranger::ranger`}
}
\value{
A tibble with three columns:
\itemize{

\item size: Number of variables used
\item ooberror: Out-of-box error of the forest
\item varimp: A list-column where each item is a data.frame with variable names and importance

}
}
\description{
Modifications to regular recursive feature elimination procedure:
\itemize{

\item Use oob prediction error as a proxy to model performance

\item Build forests on samples of data and average them

}
}
\examples{
temp <- forest_rfe(iris, "Species")
temp
}
