% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data}
\alias{generate_data}
\title{Generate mock coding/billing data frame}
\usage{
generate_data(rows = 100, count_days = FALSE, ...)
}
\arguments{
\item{rows}{\verb{[integerish]} rows number of rows to generate; default is \code{100}}

\item{count_days}{\verb{[logical]} add columns for days between events; default is \code{FALSE}}

\item{...}{\verb{[dots]} additional arguments}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate mock coding/billing data frame
}
\examples{
generate_data(rows = 5)

}
