% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parbx.R
\name{parbx_ex}
\alias{parbx_ex}
\title{Generate Mock PARBx data}
\usage{
parbx_ex(payer, ...)
}
\arguments{
\item{payer}{\verb{[character]} vector of payer names}

\item{...}{\verb{[dots]} additional arguments}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate Mock PARBx data
}
\examples{
# Every name generates 60 rows of data
parbx_ex(
  c("Cigna", "BCBS", "Anthem",
    "Humana", "UHC", "Medicare",
    "Coventry", "Medicaid", "Centene",
    "CVSHealth"))

}
