% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{predict_depth_observations.randomForest}
\alias{predict_depth_observations.randomForest}
\title{predict_depth_observations.randomForest}
\usage{
\method{predict_depth_observations}{randomForest}(object, data,
  nproc = 1)
}
\arguments{
\item{object}{A tree ensemble model}

\item{data}{data}

\item{nproc}{number of cores to use for parallel processing}
}
\description{
predict_depth_observations.randomForest
}
