% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalNodes.R
\name{predict_dissimilarity_observations_terminalNodes}
\alias{predict_dissimilarity_observations_terminalNodes}
\title{predict_dissimilarity_observations_terminalNodes}
\usage{
predict_dissimilarity_observations_terminalNodes(object, newdata, nproc)
}
\arguments{
\item{object}{object}

\item{newdata}{newdata}

\item{nproc}{number of parallel processes to use}
}
\description{
predict_dissimilarity_observations_terminalNodes
}
