% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_compute.R
\name{predict_dissimilarity_trees_terminalNodes}
\alias{predict_dissimilarity_trees_terminalNodes}
\title{predict_dissimilarity_trees_terminalNodes}
\usage{
predict_dissimilarity_trees_terminalNodes(object, newdata)
}
\description{
predict_dissimilarity_trees_terminalNodes
}
