% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalNodes.R
\name{predict_outlyingness_observations_terminalNodes}
\alias{predict_outlyingness_observations_terminalNodes}
\title{predict_outlyingness_observations_terminalNodes}
\usage{
predict_outlyingness_observations_terminalNodes(object, newdata, classes)
}
\arguments{
\item{object}{object}

\item{newdata}{newdata}

\item{classes}{(a factor) classes}
}
\description{
predict_outlyingness_observations_terminalNodes
}
