% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{proximity}
\alias{proximity}
\alias{similarity}
\title{Proximity matrix from unsupervised tree ensemble}
\usage{
proximity(dataset, method = "synthetic", predictMethod = "terminalNodes",
  ...)
}
\arguments{
\item{dataset}{A dataframe}

\item{method}{(string) Method to build the tree ensemble. These are
supported: 'synthetic'}

\item{predictMethod}{Method to predict. These are supported: 'terminalNodes'}

\item{...}{Arguments for the tree ensembler}
}
\value{
A object of class 'dist/simil'
}
\description{
Compute 'dist/simil' object from unsupervised tree ensemble
}
\details{
If method is 'synthetic', a tree ensemble is grown to seperate
  actual data from synthetic data using 'synthetic_forest' function. If
  predictMethod is 'terminalNodes', similarity matrix is computed from
  cooccurance in terminal nodes using 'forage' function.
}
\examples{
dm <- proximity(iris[, 1:4])
attr(dm, "Size")
}
