% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_claims.R
\name{summarise_claims}
\alias{summarise_claims}
\title{Summarise mock coding/billing data frame}
\usage{
summarise_claims(df, vars = c(dplyr::starts_with("days_"), dar))
}
\arguments{
\item{df}{\verb{[data.frame]} data frame generated from \code{mock_claims() |> prep_claims()}}

\item{vars}{\verb{[character]} variables to summarise, e.g. \code{c(days_service, days_reconciliation)}}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Summarise mock coding/billing data frame
}
\examples{
x <- mock_claims(rows = 500) |> prep_claims()

summarise_claims(x)

x |>
  dplyr::group_by(
    year = ymd::year(date_service),
    month = ymd::month(date_service)) |>
  summarise_claims()

}
